requirements.txt
requests
colorama
googlesearch-python
dankware